Add AlbumInfoForPostDto and SongResponseForPostDto interfaces
The code changes include the addition of two new interfaces, AlbumInfoForPostDto and SongResponseForPostDto, which are used in the post feature. These interfaces define the structure of the data objects used for album information and song responses in posts.

Based on the recent user commits and repository commits, it seems that this commit is part of a larger refactoring and feature development process.