Remove unused 'lastname' field from SignupRequest interface
The code changes remove the 'lastname' field from the SignupRequest interface in the 'src/features/auth/register/interfaces/SignupRequest.ts' file. This field is no longer needed and has been removed to simplify the interface.

Based on the recent user commits and repository commits, it seems that this change is part of a larger refactoring process.